// src/api/model/createPaymentDtoPaymentMethod.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * accounting
 * Accounting API description
 * OpenAPI spec version: 1.0
 */

export type CreatePaymentDtoPaymentMethod = typeof CreatePaymentDtoPaymentMethod[keyof typeof CreatePaymentDtoPaymentMethod];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreatePaymentDtoPaymentMethod = {
  BANK_TRANSFER: 'BANK_TRANSFER',
  CASH: 'CASH',
  CARD: 'CARD',
  OTHER: 'OTHER',
} as const;